Conflicts:
- SeQuant/domain/mbpt/op.cpp: master dropped the unused `this` capture from
  the OpMaker slot lambdas; keep our make_tensor-based lambdas without it.
- tests/unit/test_eval_btas.cpp: both sides appended a test case
  (eval_signed_leaf_phase_btas here, eval_dot_complex_btas on master); keep
  both.

+++ SeQuant/core/tensor_network/v1.cpp
@@ -720,7 +720,7 @@ TensorNetworkV1::GraphData TensorNetworkV1::make_bliss_graph(
   });
 
   // allocate graph
-  graph = std::make_shared<bliss::Graph>(nv);
+  graph = std::make_shared<bliss::Graph>(static_cast<unsigned int>(nv));
 
   // add edges
   // - each index's degree <= 2 + # of protoindex terminals
